server/requirements.txt
celery==5.2.7
Flask==2.1.3
Flask_Cors==3.0.10
gevent==21.12.0
joblib==1.1.0
onnx==1.12.0
pandas==1.4.2
protobuf==4.21.4
pypmml==0.9.16
sklearn==1.0.2
pillow==9.2.0
